SANCHAY JOSHI(2) :: data type error

+++ author-SANCHAY JOSHI(2).c
@@ -2,8 +2,8 @@
 
 int main()
 {
-	int var = 2147483648;
+	unsigned long var = 2147483648;
 	printf("Sanchay\n\n");
-	printf("Value of var = %d\n",var);
+	printf("Value of var = %lu\n",var);
 	return 0;
 }
